Feature/packet

+++ CppServer/CppServer/ChatServer.cpp
@@ -1,6 +1,7 @@
 #include "ChatServer.h"
 
-void ChatServer::Run(UINT portNum, UINT16 maxClient) {
+void ChatServer::Run(UINT16 portNum, UINT16 maxClient) {
+
 	processThread = std::thread([this]() {ProcessPacket(); });
 
 	InitSocket(portNum, maxClient);
